backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
redis
neo4j
pymongo
pandas==2.2.2
passlib
python-jose
python-dotenv
